refactor: abstract mojo with ivyEngine access; excluding compile params
- we will not necessarily need any of the standard compile params for
openApi codegen
